Flat file structure using .
